+++ testdata/tree/integration_tests/src/lib.rs
@@ -1,0 +1,3 @@
+pub fn double(n: u32) -> u32 {
+    2 * n
+}