+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fixed: Don't try to mutate functions within test targets, e.g. within `tests/**/*.rs`.
+
 - New: `missed.txt`, `caught.txt`, `timeout.txt` and `unviable.txt` files are written in to the output directory to make results easier to review later.
 
 - New: `--output` creates the containing directory if it does not exist.